Move Flow type definitions into index
